completed team activity
