+++ src/css/style.css
@@ -185,7 +185,7 @@ button {
 }
 
 .product-card__savings-percent {
-  color: #FFFFFF;
+  color: #ffffff;
   background-color: #525b0f;
   font-size: 14px;
   line-height: 1;
