+++ src/public/partials/footer.html
@@ -1,0 +1,1 @@
+<footer>&copy;NOT a real business</footer>
